Module Rcontext

Per-request contexts

module Ctx : sig ... end

Main type

include module type of struct include Ctx end

Main type definition

type 'a key = 'a Hmap.key

Key to access a certain value in the context

type t = Hmap.t

A request context. This is simply a hmap, and standard keys into it are provided in the rest of the library.

val empty : t

Empty context

Deadlines

Depends on mtime.

val k_deadline : Mtime.t Hmap.key

Key to access a deadline from inside the request context. This can be useful for systems concerned with timeout or cancellation of a request after a certain amount of time has passed.

val deadline : Ctx.t -> Mtime.t option

val deadline_exn : Ctx.t -> Mtime.t

val with_deadline : Mtime.t -> Ctx.t -> Ctx.t

val expired : Ctx.t -> bool

Has the deadline been reached? This uses Mtime_clock.now to compare it to the stored deadline. If there is no deadline then this returns false.

Start time for a request or task

Depends on mtime.

val k_start_time : Mtime.t Hmap.key

Instant at which an action was started, a request was received, etc. This can be useful for middlewares related to timing, tracing, latency statistics, etc.

val start_time : Ctx.t -> Mtime.t option

val start_time_exn : Ctx.t -> Mtime.t

val with_start_time : Mtime.t -> Ctx.t -> Ctx.t

val with_start_time_now : Ctx.t -> Ctx.t

Ambient rcontext via thread local storage

module Ambient_hmap : sig ... end

A way to store a rcontext in thread-local-storage. This depends on the thread-local-storage library.