Module Rcontext.Ctx

Main type

Main type definition

type 'a key = 'a Hmap.key

Key to access a certain value in the context

type t = Hmap.t

A request context. This is simply a hmap, and standard keys into it are provided in the rest of the library.

val empty : t

Empty context

val add : 'a key -> 'a -> t -> t

val get : 'a key -> t -> 'a option

val get_exn : 'a key -> t -> 'a