Module Rcontext

Per-request contexts

type 'a key = 'a Hmap.key

Key to access a certain value in the context

type t = {

values : Hmap.t;

}

A request context

val empty : t

Empty context

val k_deadline : Mtime.t Hmap.key

Key to access a deadline from inside the request context. This can be useful for systems concerned with timeout or cancellation of a request after a certain amount of time has passed.

val deadline : Rcontext__.Ctx.t -> Mtime.t option

val deadline_exn : Rcontext__.Ctx.t -> Mtime.t

val with_deadline : Mtime.t -> Rcontext__.Ctx.t -> Rcontext__.Ctx.t

val expired : Rcontext__.Ctx.t -> bool

Has the deadline been reached? This uses Mtime_clock.now to compare it to the stored deadline. If there is no deadline then this returns false.

val k_start_time : Mtime.t Hmap.key

Instant at which an action was started, a request was received, etc. This can be useful for middlewares related to timing, tracing, latency statistics, etc.

val start_time : Rcontext__.Ctx.t -> Mtime.t option

val start_time_exn : Rcontext__.Ctx.t -> Mtime.t

val with_start_time : Mtime.t -> Rcontext__.Ctx.t -> Rcontext__.Ctx.t

val with_start_time_now : Rcontext__.Ctx.t -> Rcontext__.Ctx.t